Put back types with their class user
